fix: four Phase 2 code quality fixes
1. Relax ProfitTester basis check: now accepts GROSS and NET, rejects
   only CEDED. Removes the 20-line CashFlowResult re-labeling workaround
   in MonteCarloUQ's no-treaty path — gross cash flows are passed directly.

2. Remove redundant .astype(np.float64) calls in improvement.py helper
   functions — source arrays are already float64 so advanced indexing
   preserves dtype; keep the type: ignore[no-any-return] for mypy.

3. Rename WholeLife._compute_net_premiums -> _compute_annual_net_premiums
   and update docstring to clarify the return is annual, not monthly
   (caller divides by 12 at line 249).

4. Fix logic error in test_di_benefit_grows_over_time: 'or' -> 'and'
   so both growth and positivity conditions must hold; fix year-1 slice
   from [1:12] to [0:12] to include month 0.

+++ src/polaris_re/assumptions/improvement.py
@@ -139,13 +139,13 @@ _CPM_B_FACTORS: np.ndarray = np.array(
 def _get_scale_aa_factors(ages: np.ndarray) -> np.ndarray:
     """Look up Scale AA improvement factors for a vector of ages."""
     capped_ages = np.clip(ages, 0, len(_SCALE_AA_FACTORS) - 1).astype(np.int32)
-    return _SCALE_AA_FACTORS[capped_ages].astype(np.float64)  # type: ignore[no-any-return]
+    return _SCALE_AA_FACTORS[capped_ages]  # type: ignore[no-any-return]
 
 
 def _get_cpm_b_factors(ages: np.ndarray) -> np.ndarray:
     """Look up CPM-B improvement factors for a vector of ages."""
     capped_ages = np.clip(ages, 0, len(_CPM_B_FACTORS) - 1).astype(np.int32)
-    return _CPM_B_FACTORS[capped_ages].astype(np.float64)  # type: ignore[no-any-return]
+    return _CPM_B_FACTORS[capped_ages]  # type: ignore[no-any-return]
 
 
 def _get_mp2020_factors_for_year(ages: np.ndarray, calendar_year: int) -> np.ndarray:
@@ -158,7 +158,7 @@ def _get_mp2020_factors_for_year(ages: np.ndarray, calendar_year: int) -> np.nda
     capped_ages = np.clip(ages, 0, 120).astype(np.int32)
     year_offset = calendar_year - _MP2020_BASE_YEAR
     year_offset = int(np.clip(year_offset, 0, _MP2020_FACTORS.shape[1] - 1))
-    return _MP2020_FACTORS[capped_ages, year_offset].astype(np.float64)  # type: ignore[no-any-return]
+    return _MP2020_FACTORS[capped_ages, year_offset]  # type: ignore[no-any-return]
 
 
 class MortalityImprovement(PolarisBaseModel):
